Add copy mechanism to ParamContainer
Add explicit copy constructor and make the class abstract through
addition of a pure virtual bind_params() method, which derived
classes can use to make a deep copy of objects. This is required
since it is not possible for the base ParamContainer class to rebind
parameters to setter/getter method in the copy. Also adds other
changes to facilitate the copy mechanism